% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epith.R
\name{epith}
\alias{epith}
\title{Title: Calculate Epidemic Threshold}
\usage{
epith(
  data,
  disease,
  latest_week,
  latest_year,
  area,
  alert_sd = 1,
  epidemic_sd = 1.65,
  onset_column = "donset"
)
}
\arguments{
\item{disease}{Disease label to add to the title of the plot}

\item{latest_week}{Column name for MMWR week.}

\item{latest_year}{Column name for year.}

\item{area}{Area label to add to the title of the plot}

\item{alert_sd}{One (1) standard deviation above the mean (default is 1)}

\item{epidemic_sd}{One (1.65) standard deviation above the mean (default is 1.65)}

\item{onset_column}{Column name for date of onset (default is "donset")}

\item{df}{A data frame of disease counts.}
}
\value{
A data frame with epidemic and alert thresholds.
}
\description{
This function calculates the alert and epidemic thresholds based on historical data.
}
